playlist options to bottom sheet
