+++ app/src/main/java/com/github/libretube/adapters/PlaylistsAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.github.libretube.R
 import com.github.libretube.api.RetrofitInstance
 import com.github.libretube.databinding.PlaylistsRowBinding
-import com.github.libretube.sheets.PlaylistOptionsDialog
+import com.github.libretube.sheets.PlaylistOptionsBottomSheet
 import com.github.libretube.extensions.TAG
 import com.github.libretube.obj.PlaylistId
 import com.github.libretube.obj.Playlists
@@ -71,13 +71,13 @@ class PlaylistsAdapter(
             }
 
             root.setOnLongClickListener {
-                val playlistOptionsDialog = PlaylistOptionsDialog(
+                val playlistOptionsDialog = PlaylistOptionsBottomSheet(
                     playlistId = playlist.id!!,
                     isOwner = true
                 )
                 playlistOptionsDialog.show(
                     childFragmentManager,
-                    PlaylistOptionsDialog::class.java.name
+                    PlaylistOptionsBottomSheet::class.java.name
                 )
                 true
             }
